added writing to DB

+++ src/actions-stats/Models/WorkflowRun.cs
@@ -4,11 +4,11 @@ namespace ActionsStats.Models;
 
 public class WorkflowRun
 {
-    public int Id { get; set; }
-    public int RunNumber { get; set; }
+    public long Id { get; set; }
+    public long RunNumber { get; set; }
     public string Org { get; set; }
     public string Repo { get; set; }
-    public int WorkflowId { get; set; }
+    public long WorkflowId { get; set; }
     public string WorkflowName { get; set; }
     public string Actor { get; set; }
     public string Branch { get; set; }
